Use rangeUtils property
