--- src/language-html/is-source-element.js
@@ -1,10 +1,0 @@
-"use strict";
-
-function isSourceElement(node, parentNode, opts) {
-  if (opts.parser === "vue") {
-    return node.tag !== "root";
-  }
-  return false;
-}
-
-module.exports = isSourceElement;
