+++ src/language-html/printer-html.js
@@ -60,7 +60,6 @@ const {
   isVueEventBindingExpression,
 } = require("./syntax-vue.js");
 const { printImgSrcset, printClassNames } = require("./syntax-attribute.js");
-const isSourceElement = require("./is-source-element.js");
 
 function embed(path, print, textToDoc, options) {
   const node = path.getValue();
@@ -1181,5 +1180,4 @@ module.exports = {
   insertPragma,
   massageAstNode: clean,
   embed,
-  isSourceElement,
 };
