+++ src/language-html/range-utils.js
@@ -1,0 +1,7 @@
+"use strict";
+
+function isVueSourceElement(node) {
+  return node.tag !== "root";
+}
+
+module.exports = { isVueSourceElement };
