+++ src/language-js/parse/utils/create-parser.js
@@ -2,6 +2,7 @@
 
 const { hasPragma } = require("../../pragma.js");
 const { locStart, locEnd } = require("../../loc.js");
+const { isSourceElement } = require("../../range-utils.js");
 
 function createParser(options) {
   options = typeof options === "function" ? { parse: options } : options;
@@ -11,6 +12,9 @@ function createParser(options) {
     hasPragma,
     locStart,
     locEnd,
+    rangeUtils: {
+      isSourceElement,
+    },
     ...options,
   };
 }
