+++ src/language-js/utils.js
@@ -1310,20 +1310,6 @@ function isObjectProperty(node) {
   );
 }
 
-const jsonSourceElements = new Set([
-  "ObjectExpression",
-  "ArrayExpression",
-  "StringLiteral",
-  "NumericLiteral",
-  "BooleanLiteral",
-  "NullLiteral",
-  "UnaryExpression",
-  "TemplateLiteral",
-]);
-function isJsonSourceElement(node) {
-  return jsonSourceElements.has(node.type);
-}
-
 module.exports = {
   getFunctionParameters,
   iterateFunctionParametersPath,
@@ -1356,7 +1342,6 @@ module.exports = {
   isFunctionOrArrowExpression,
   isGetterOrSetter,
   isJestEachTemplateLiteral,
-  isJsonSourceElement,
   isJsxNode,
   isLiteral,
   isLongCurriedCallExpression,
